drm/i915: MOCS settings for gen9 to match media driver
This patch attempts to match MOCS settings as they are expected by
this media driver: https://github.com/intel/media-driver. Driver
can work with the previos default settings, but it is supposed
it should work faster with these settings. Behavior is
differentiated by HAVE_MESA_MOCS compile level define. This patch
is NOT intended for the integration into the i915, but to try out
these settings and find out whether or not they are beneficial.

4.19.0
    ported to 4.19.0 for KBL-G.[Owen Zhang]

Change-Id: I427f7f92cbd2f23f30ca5a1af43eedd229cf41ff
Cc: David Weinehall <david.weinehall@linux.intel.com>
Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>
Signed-off-by: Owen Zhang <owen.zhang@intel.com>