drm/i915: Record both min/max eu_per_subslice in sseu_dev_info
When we query the available eu on each subslice, we currently only
report the max. It would also be useful to report the minimum found as
well.

When we set RPCS (power gating over the EU), we can also specify both
the min and max number of eu to configure on each slice; currently we
just set it to a single value, but the flexibility may be beneficial in
future.

Change-Id: I96dd4e5ce9d9ecf147ac5baca992d51884f8aff4
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>
Reviewed-by: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Rfc: https://patchwork.freedesktop.org/series/29715/#rev2